ai-voice-agent-backend/requirements.txt
fastapi==0.116.2
uvicorn[standard]==0.35.0
python-multipart==0.0.6
python-dotenv==1.1.1
supabase==2.19.0
httpx==0.28.1
pydantic==2.11.9
pydantic-settings==2.10.1
